Make verification for "XYZ"-type destinations more forgiving.
In testing, I have found an example of a PDF for which "XYZ"-type destinations (as computed by the call to `this.pdfDocument.getDestination()` at web/pdf_link_service.js:168) come back as arrays of length 4, not 5. In other words, there is no "scale" entry.

Moreover, it seems that there is only one place in the code where such destinations are actually _interpreted_, namely in web/base_viewer.js:scrollPageIntoView, and there it looks as if the code is entirely tolerant of a missing fifth entry. So I don't think the verification needs to be so strict.